fix(tools): address PR #144 review — content gate is necessary, not sufficient
- Reframe the card `content` attribute as necessary-but-not-sufficient for
  export: the adapter sets it via the lenient supports_data_export() while
  /api/v1/contents gates on the stricter export_safe(), so a content-bearing
  card can still 403. Updated the output-schema .describe(), the tako_contents
  description/header, and the llms-full.txt precondition; one-liner surfaces
  already stated only the necessary direction.
- Stop the 403 message asserting the card "carried no `content` attribute" —
  a content-present card is exactly what reaches that branch surprisingly.
- Route 403/404 upstream bodies through extractErrorDetail() instead of a raw
  slice, so an edge/WAF HTML block page never floods the model text; moved
  extractErrorDetail from mcp.ts to django.ts to avoid an import cycle, which
  also reunites the stranded djangoErrorToToolResult JSDoc with its function.
- New test: unstructured 403 body is omitted from the model-visible message.
