+++ llms-full.txt
@@ -66,7 +66,7 @@ Parameters:
 ### tako_contents
 Fetch the actual rows or data behind a result URL — the step after `tako_search`/`tako_answer` when you need the real underlying data to compute over, quote, or reason about (a search result carries only metadata and a chart, not its data). A Tako card URL returns that card's data (CSV by default, or JSON); any other URL — a web result — returns that page's extracted full text. One URL per call.
 
-**Exportability precondition:** only a Tako card whose search/answer result carried a `content` attribute (non-null) is exportable — `content` is the card's export descriptor. A card that came back with `content` missing or null has NO exportable data; calling `tako_contents` on it fails, so don't — use that card's title, inline preview, and chart instead. Web result URLs are exempt from this gate (they resolve to page text).
+**Exportability precondition:** only a Tako card whose search/answer result carried a `content` attribute (non-null) is exportable — `content` is the card's export descriptor. A card that came back with `content` missing or null has NO exportable data; calling `tako_contents` on it fails, so don't — use that card's title, inline preview, and chart instead. Presence is required but not a guarantee: the export gate can still refuse a rare card (a self-correcting 403 — fall back to its preview/chart, don't retry). Web result URLs are exempt from this gate (they resolve to page text).
 
 Parameters:
 - `url` (string, required): The result URL to fetch — a Tako card's `webpage_url`, or a web-result URL
